pyproject.toml
[tool.poetry]
name = "graphrag"
version = "0.1.0"
description = "Graph-Enhanced RAG system combining vector search, knowledge graphs, and SQL for hybrid retrieval"
authors = ["HUSKI Team"]
readme = "README.md"
packages = [{include = "graphrag", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.1.0"
langchain-openai = "^0.0.5"
langchain-google-genai = "^2.0.0"
langchain-groq = "^0.1.9"
langchain-community = "^0.0.20"
chromadb = "^0.4.22"
neo4j = "^5.15.0"
openai = "^1.10.0"
google-generativeai = "^0.8.0"
groq = "^0.11.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
sentence-transformers = "^2.3.1"
tiktoken = "^0.5.2"
sqlparse = "^0.4.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.3"
black = "^23.12.1"
ruff = "^0.1.9"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
